MC Refactor
Moved all mc source files into a new package called 'fmcb' and replaced any instances of 'main.java' with 'fmcb'. This is in line with how gradle treats the project, instead of how eclipse does by default.
