--- Minecraft/FMCB/src/main/java/FMCB.java
@@ -1,34 +1,0 @@
-package main.java;
-
-import main.java.init.ModBlocks;
-import main.java.init.ModItems;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.Mod.EventHandler;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-
-@Mod(modid = "fmcb", useMetadata = true)
-public class FMCB {
-
-	// Needed?
-	@Mod.Instance
-	public static FMCB instance;
-
-	@EventHandler
-	public void preInit(FMLPreInitializationEvent event) {
-		ModItems.init();
-		ModBlocks.init();
-	}
-
-	@EventHandler
-	public void init(FMLInitializationEvent event) {
-		
-	}
-	
-	@EventHandler
-	public void postInit(FMLPostInitializationEvent event) {
-
-	}
-	
-}
