+++ Minecraft/FMCB/src/main/java/fmcb/FMCB.java
@@ -1,0 +1,34 @@
+package fmcb;
+
+import fmcb.init.ModBlocks;
+import fmcb.init.ModItems;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.Mod.EventHandler;
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+@Mod(modid = "fmcb", useMetadata = true)
+public class FMCB {
+
+	// Needed?
+	@Mod.Instance
+	public static FMCB instance;
+
+	@EventHandler
+	public void preInit(FMLPreInitializationEvent event) {
+		ModItems.init();
+		ModBlocks.init();
+	}
+
+	@EventHandler
+	public void init(FMLInitializationEvent event) {
+		
+	}
+	
+	@EventHandler
+	public void postInit(FMLPostInitializationEvent event) {
+
+	}
+	
+}
