+++ Minecraft/FMCB/src/main/java/fmcb/block/BlockGeneral.java
@@ -1,0 +1,22 @@
+package fmcb.block;
+
+import fmcb.init.ModBlocks;
+import net.minecraft.block.Block;
+import net.minecraft.block.material.Material;
+import net.minecraft.creativetab.CreativeTabs;
+
+public class BlockGeneral extends Block {
+	
+	public BlockGeneral(Material material, String name) {
+		super(material);
+
+		setUnlocalizedName(name);
+		setRegistryName(name);
+		
+		setCreativeTab(CreativeTabs.FOOD);
+		
+		ModBlocks.blocks.add(this);
+	}
+
+
+}
