+++ Minecraft/FMCB/src/main/java/fmcb/side/client/ClientProxy.java
@@ -1,0 +1,5 @@
+package fmcb.side.client;
+
+public class ClientProxy {
+
+}
