--- Minecraft/FMCB/src/main/java/side/client/ClientProxy.java
@@ -1,5 +1,0 @@
-package main.java.side.client;
-
-public class ClientProxy {
-
-}
