--- Minecraft/FMCB/src/main/java/side/server/ServerProxy.java
@@ -1,25 +1,0 @@
-package main.java.side.server;
-
-import net.minecraftforge.fml.common.Mod.EventHandler;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-
-public class ServerProxy {
-
-	@EventHandler
-	public void preInit(FMLPreInitializationEvent event) {
-		
-	}
-
-	@EventHandler
-	public void init(FMLInitializationEvent event) {
-		
-	}
-	
-	@EventHandler
-	public void postInit(FMLPostInitializationEvent event) {
-		
-	}
-	
-}
